Product types - update schema and add CRUD operations

+++ src/models/product-collections/product-collections.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import {
   CreateProductCollectionDto,
   UpdateProductCollectionDto,
@@ -8,12 +8,14 @@ import { InjectModel } from '@nestjs/mongoose';
 import { ProductCollection } from '@/models/product-collections/schemas/product-collection.schema';
 import { Model } from 'mongoose';
 import { ResponseProductCollectionDto } from '@/models/product-collections/dto/response-product-collection.dto';
+import { ProductTypesService } from '@/models/product-types/product-types.service';
 
 @Injectable()
 export class ProductCollectionsService {
   constructor(
     @InjectModel(ProductCollection.name)
     private productCollectionModel: Model<ProductCollection>,
+    private productTypeService: ProductTypesService,
   ) {}
   async create(
     createProductCollectionDto: CreateProductCollectionDto,
@@ -76,6 +78,14 @@ export class ProductCollectionsService {
   }
 
   async remove(id: string) {
+    const productTypeCount =
+      await this.productTypeService.getProductTypeCountForCollection(id);
+
+    if (productTypeCount > 0)
+      throw new Error(
+        `Cannot delete product collection. This collection has ${productTypeCount} product types assigned to it.`,
+      );
+
     const deletedProductCollection = await this.productCollectionModel
       .findByIdAndDelete(id)
       .lean()
