+++ src/models/product-types/product-types.controller.ts
@@ -5,7 +5,9 @@ import {
   Body,
   Patch,
   Param,
-  Delete, HttpException, HttpStatus,
+  Delete,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import {
   CreateProductTypeDto,
@@ -14,44 +16,93 @@ import {
 import { User } from '@/common/decorators';
 import { ProductTypesService } from './product-types.service';
 import { MutationSuccessResponse, RequestUser } from '@/interfaces';
+import mongoose from 'mongoose';
+import { ResponseProductTypeDto } from '@/models/product-types/dto/response-product-type.dto';
 
 @Controller('product-types')
 export class ProductTypesController {
   constructor(private readonly productTypesService: ProductTypesService) {}
 
   @Post()
-  async create(@Body() createProductTypeDto: CreateProductTypeDto, @User() user: RequestUser): Promise<MutationSuccessResponse> {
-    try{
-      const newProductType = await this.productTypesService.create(createProductTypeDto, user.fullName);
+  async create(
+    @Body() createProductTypeDto: CreateProductTypeDto,
+    @User() user: RequestUser,
+  ): Promise<MutationSuccessResponse> {
+    try {
+      const newProductType = await this.productTypesService.create(
+        createProductTypeDto,
+        user.fullName,
+      );
       return {
         id: newProductType._id.toString(),
         message: 'Product Type created successfully',
-      }
+      };
     } catch (err) {
       throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
     }
   }
 
   @Get()
-  findAll() {
-    return this.productTypesService.findAll();
+  async findAll(): Promise<ResponseProductTypeDto[]> {
+    try {
+      return await this.productTypesService.findAll();
+    } catch (err) {
+      throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
+    }
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.productTypesService.findOne(id);
+  async findOne(@Param('id') id: string) {
+    const isIdValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isIdValid)
+      throw new HttpException('Product Type not found', HttpStatus.BAD_REQUEST);
+
+    try {
+      return await this.productTypesService.findOne(id);
+    } catch (err) {
+      throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
+    }
   }
 
   @Patch(':id')
-  update(
+  async update(
     @Param('id') id: string,
     @Body() updateProductTypeDto: UpdateProductTypeDto,
+    @User() user: RequestUser,
   ) {
-    return this.productTypesService.update(id, updateProductTypeDto);
+    const isIdValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isIdValid)
+      throw new HttpException('Product Type not found', HttpStatus.BAD_REQUEST);
+
+    try {
+      await this.productTypesService.update(
+        id,
+        updateProductTypeDto,
+        user.fullName,
+      );
+
+      return {
+        id,
+        message: 'Product Type updated successfully',
+      };
+    } catch (err) {
+      throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
+    }
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.productTypesService.remove(id);
+  async remove(@Param('id') id: string): Promise<MutationSuccessResponse> {
+    const isIdValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isIdValid)
+      throw new HttpException('Product Type not found', HttpStatus.BAD_REQUEST);
+
+    try {
+      await this.productTypesService.remove(id);
+      return {
+        message: 'Product Type deleted successfully',
+      };
+    } catch (err) {
+      throw new HttpException(err.message, HttpStatus.BAD_REQUEST);
+    }
   }
 }
